refactor: lib/iBooking => lib/integration/sit
