+++ components/modals/Agenda/Agenda.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography, useTheme } from "@mui/material";
 import AgendaClassItem, { AgendaClass } from "./AgendaClassItem";
 import React from "react";
-import { classConfigRecurrentId } from "../../../lib/iBooking";
+import { classConfigRecurrentId } from "../../../lib/integration/sit";
 import { SitClass } from "../../../types/integration/sit";
 import { ClassConfig } from "../../../types/rezervo";
 import { WEEKDAY_NUMBER_TO_NAME } from "../../../utils/timeUtils";
