+++ components/modals/Agenda/AgendaModal.tsx
@@ -1,5 +1,5 @@
 import { Modal } from "@mui/material";
-import { classConfigRecurrentId, sitClassRecurrentId } from "../../../lib/iBooking";
+import { classConfigRecurrentId, sitClassRecurrentId } from "../../../lib/integration/sit";
 import React, { Dispatch, SetStateAction } from "react";
 import Agenda from "./Agenda";
 import { UserConfig } from "../../../types/rezervo";
