+++ components/modals/ClassInfo/ClassInfoModal.tsx
@@ -1,6 +1,6 @@
 import { Modal } from "@mui/material";
 import ClassInfo from "./ClassInfo";
-import { sitClassRecurrentId } from "../../../lib/iBooking";
+import { sitClassRecurrentId } from "../../../lib/integration/sit";
 import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex } from "../../../types/rezervo";
 import React, { Dispatch, SetStateAction } from "react";
 import { SitClass } from "../../../types/integration/sit";
