+++ components/schedule/DaySchedule.tsx
@@ -1,6 +1,6 @@
 import { Box, Chip, Typography, useTheme } from "@mui/material";
 import ClassCard from "./class/ClassCard";
-import { sitClassRecurrentId } from "../../lib/iBooking";
+import { sitClassRecurrentId } from "../../lib/integration/sit";
 import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex } from "../../types/rezervo";
 import React from "react";
 import { SitClass, SitDaySchedule } from "../../types/integration/sit";
